Fix building packages on Python 3.10.
Two parts of the build process, `build-i18n.py` and `build-media.py`,
trigger the `collections` import issues on Python 3.10. This wasn't a
problem for releases before due to our build process favoring building
multi-version wheels with Python 2.7 and due to eggs not being built for
Python 3.x, but it's going to be a problem going forward as that logic
has changed.

We now apply all the Django Evolution patches, which includes the
collections patch, to Django before Django itself is initialized. This
ensures we can build on all supported versions of Python.

Testing Done:
Successfully built packages on Python 3.10.

Reviewed at https://reviews.reviewboard.org/r/12156/