Update Python and dependency versions for 5.0.x.
We're going to be dropping support for Python 2.x for the 5.0 release.
We already have made some changes to the code that affect this, but our
version specifiers and dependencies still had 2.x stuff in there. This
change updates everything so that we're Python 3.7+, and should make it
possible to run `setup.py --all-pyvers` successfully again.

Testing Done:
Ran setup.py and unit tests.

Reviewed at https://reviews.reviewboard.org/r/12205/