Add fake SCMTool for VersionVault, and fix display of advert for tools.
With the addition of the new Power Pack based VersionVault
implementation, we need an entry to advertise it for users who don't
have Power Pack installed. While I was doing this, I noticed that we
weren't properly showing the advert blurb for fake tools, only for fake
hosting services, so I added in the implementation for that.

Testing Done:
Set up a server that didn't have Power Pack installed. Flipped through
the various SCMTools and saw a Power Pack blurb for tfs, tfs-git, and
versionvault, and that the submission buttons were properly disabled.

Reviewed at https://reviews.reviewboard.org/r/12207/