Add documentation on using Cliosoft SOS with Review Board.
This covers the requirements for SOS on Review Board, how to set up a
SOS repository, and how to configure RBTools.

This will be the primary documentation used for SOS on the server side.
The Power Pack documentation will provide an introduction and some
details, but will otherwise point to this guide.

Testing Done:
Built the docs. Checked for bad links, build errors, and spelling
errors.

Reviewed at https://reviews.reviewboard.org/r/12265/