Add docstrings with version information to all evolutions.
Review Board has had a great many database schema changes over its
lifetime, and the more we add, the harder it is to determine when a
particular change was made.

This poses a problem when trying to sort out failed evolution upgrades,
as sometimes an upgrade is attempted and rolled back with evolution
information recorded that breaks future attempts at a new upgrade. When
this happens, it takes a lot of effort to figure out what happened.

This change updates all evolutions in the codebase to have a docstring
that specifies the version in which the evolution was added. This rounds
up to the first public release (so no alphas/betas).

All upgrades are also tracked in Notion, for easy reference:

https://reviewboard.notion.site/Database-Schema-Changes-38702981914e48a3b32f2a3dc5243a5c

While here, I found a stray evolution that was never used (its contents
were incorporated into another evolution). That's been deleted.

Testing Done:
Went through and verified the versions of each of these.

Reviewed at https://reviews.reviewboard.org/r/12290/