Fix a missing default and missing recorded migrations with oauth2_provider.
This fixes up some issues with the oauth2_provider upgrade.

First, we now set an explicit default value for `Grant.claims`, which is
needed on everything but MySQL and was missing in my previous change.

We also now manage the list of recorded migrations for oauth2_provider.
Post-upgrade, we do another clear of the list of recorded migrations and
then we populate it with the full list that ships with oauth2_provider
1.6.3. This avoids potential dependency issues with migrations, and it
avoids the Django system checks complaining about missing migrations
that still need to be applied.

Testing Done:
Tested a new round of upgrades. Verified that Django's system checks
no longer saw missing migrations.

Reviewed at https://reviews.reviewboard.org/r/12293/