Fix building docs for Review Board 5.
This updates the initial data fixtures and the `webapidocs` extension to
enable building documentation.

The fixtures include new fields required for the supported version of
OAuth2.

The extension updates just remove usage of `six`.

We also now depend on the 4.x range for `sphinx` and the 1.x range for
`beanbag-docutils`, using the `~=` version specifier.

Testing Done:
Successfully built the manual and the release notes.

Reviewed at https://reviews.reviewboard.org/r/12321/