Upgrade Haystack to 3.2.x.
This upgrades to the latest stable version of Haystack, 3.2.x, which is
compatible with Django 3.2 and 4.0.

There isn't a lot behind the scenes of this release, but it does address
some deprecation warnings that were shown in 4.0. There are a few bug
fixes behind the scenes for backends, but nothing stood out as being
especially notable for our uses. Mostly, this is a forward-looking
version bump.

Testing Done:
Upgraded and verified this took care of deprecation warnings.

All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12325/