Update Profile unit tests to use the new assertQueries() method.
`Profile` has recently had a number of optimizations, and there may be
more to come. To help ensure optimizations are having an effect on query
complexity, this change updates all the tests to use the new
`assertQueries()` in Djblets instead of `assertNumQueries()`.

This lets us check the specifics of the queries, including details on
JOINs, annotations, fields that are queried, and more. Further
improvements to this functionality will be more easily verified through
unit test changes and in code reviews.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12348/